Build and publish Docker image
This commit triggers a GitHub Actions workflow to build and publish a Docker image for the 'dev' branch. The Docker image will be pushed to the GitHub Container Registry (ghcr.io) and can be used for deployment to the hosting environment. This workflow ensures that the latest changes are containerized and ready for deployment..